Added warning when the result of an assignment expression is used
The original cog compiler did not implement assignment expressions
correctly. For example, the following statement:

x = y = z = 5;

is equivalent to:

z = 5;
x = y;

It is therefore recommended that cog scripts do not rely upon the
correct behavior of using an assignment expression as an rvalue.

The Gorc cog interpreter implements this language feature correctly.