Hacked together some texture override code for testing
Modified the material utility to accept an optional colormap. If
specified, the colormap is used for dumping the material cels and mipmap
levels to disk. Previously, material could only output an indexed
bitmap.

Modified the material loader to look for an override png on disk.

This is a relatively quick and dirty hack I used to test Gorc with some
upsampled textures. Texture override is an obviously desirable feature,
so I am pushing these changes for future use.